Nexus destination changes
Removing the ability to use closures as destinations because it breaks
the process flow. This might be added back in a future release, but for
now using strictly classes is preferred so that the power of the
container can be utilized.
Swapping out phpunit for phpspec because I prefer phpspec for unit
tests and was able to create some work arounds for tests.